Refactor ProductCard imports and create new ProductCard components for better organization
