+++ app/components/cart/cart-best-sellers.tsx
@@ -6,7 +6,7 @@ import clsx from "clsx";
 import { useEffect, useId } from "react";
 import { useFetcher } from "react-router";
 import type { ProductCardFragment } from "storefront-api.generated";
-import { ProductCard } from "~/components/product/product-card";
+import { ProductCard } from "~/components/product-card";
 import { Skeleton } from "~/components/skeleton";
 import { usePrefixPathWithLocale } from "~/hooks/use-prefix-path-with-locale";
