+++ app/components/product-card/product-card-options.tsx
@@ -7,11 +7,11 @@ import type {
 } from "storefront-api.generated";
 import { Button } from "~/components/button";
 import { Link } from "~/components/link";
+import { OPTIONS_AS_SWATCH } from "~/components/product/product-option-values";
 import { RevealUnderline } from "~/components/reveal-underline";
 import { Tooltip, TooltipContent, TooltipTrigger } from "~/components/tooltip";
 import { cn } from "~/utils/cn";
 import { isLightColor, isValidColor } from "~/utils/misc";
-import { OPTIONS_AS_SWATCH } from "./product-option-values";
 
 export function ProductCardOptions({
   product,
