+++ app/components/product-card/quick-shop.tsx
@@ -18,13 +18,13 @@ import type {
 import { Button } from "~/components/button";
 import { Link } from "~/components/link";
 import { AddToCartButton } from "~/components/product/add-to-cart-button";
+import { ProductBadges } from "~/components/product/badges";
 import { ProductMedia } from "~/components/product/product-media";
 import { Quantity } from "~/components/product/quantity";
+import { VariantPrices } from "~/components/product/variant-prices";
+import { VariantSelector } from "~/components/product/variant-selector";
 import { Skeleton } from "~/components/skeleton";
 import JudgemeStarsRating from "~/sections/main-product/judgeme-stars-rating";
-import { ProductBadges } from "./badges";
-import { VariantPrices } from "./variant-prices";
-import { VariantSelector } from "./variant-selector";
 
 interface QuickViewData {
   product: NonNullable<ProductQuery["product"]>;
